requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
scikit-learn==1.4.2
scipy==1.13.0
nltk==3.8.1
flask==3.1.2
# Environment variables 
python-dotenv==1.0.1
notebook
ipykernel

